backend/requirements.txt
pipecat-ai[webrtc,openai,deepgram,cartesia]>=0.0.100
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-dotenv>=1.0.0
loguru>=0.7.0
